fix(pdf): add missing newline and update math block to use align* environment
This change fixes a bug where the generated LaTeX code for math blocks did not include a newline at the end of the equation. Additionally, the code has been updated to use the align* environment instead of displaystyle. This allows for multiple equations to be aligned and for line breaks to be used with \\, but also means that empty lines are not allowed. The reference documentation has been updated to reflect these changes.