fix(pdf): reorder registry headings
This commit reorders the registry headings, specifically moving the citation heading to after the figures heading. This ensures that the TOF is generated in the correct order. The TODO comment for adding other registry headings was removed as it is no longer necessary. Also, this commit updates the version number in globals/version.go from 2.7.9 to 2.7.10.